refactor: extract applicableFlags helper and add field-parameter guard
Extract triplicated applicableFlags computation from dataset-settings.tsx into getApplicableFlags() in dataset-settings-utils.ts.

Add defensive guard in buildProcessingPlan to skip field-parameter role fields that escape parameter group capture, removing an unsafe type cast.

Ref: #238
